device: Build kernel in-line with device tree and remove unused variants
Hammerhead is going to use a kernel other than Google's one. That kernel
has another update cycle, so build it in-line with the device tree to
awlways ensure that we have the latest iteration of it.

Also, this commit removes a support for SveLTE and FP kernels, that are
never used anyway. Fingerprint policies and configurations are killed too
as production variant of hammerhead does not wield a fingerprint sensor.

Note that a support for building kernel in-line with device tree must be
implemented in platform/build.

Change-Id: Ia4ae0c51976fdcbb43221fc365764388adfe7652
Signed-off-by: Alex Saiko <solcmdr@gmail.com>